MainHandler contstructor takes working-directory
The MainHandler now takes in the server-working-directory, this is so
that it can pass it along to the handler functions that process a
request. This is done so that handlers that need to do file access can
have knowledge of the root working dir.

A side effect is that now every hanlder now takes in a response and the
working directory, even if they don't need either one.